LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-grid/subgrid/align-self-baseline-with-subgrid-mbp-expected.txt

X
X

FAIL .first-baseline 1 assert_equals:
<div data-offset-y="31" class="first-baseline">X</div>
offsetTop expected 31 but got 16
FAIL .first-baseline 2 assert_equals:
<div data-offset-y="31" class="first-baseline">X</div>
offsetTop expected 31 but got 16
